chore(nlu): rm jest from nlu dependencies (#4880)
