+++ src/nlu/package.json
@@ -102,8 +102,6 @@
     "@types/express": "^4.16.0",
     "@types/fs-extra": "^5.0.4",
     "@types/ioredis": "^4.0.10",
-    "@types/jest": "^26.0.22",
-    "@types/jest-environment-puppeteer": "^4.3.1",
     "@types/joi": "^13.4.5",
     "@types/jsonwebtoken": "^8.3.7",
     "@types/lodash": "^4.14.116",
@@ -147,11 +145,6 @@
     "gulp-typedoc": "^2.2.1",
     "gulp-typescript": "^6.0.0-alpha.1",
     "husky": "^4.2.5",
-    "jest": "^24.9.0",
-    "jest-circus": "^24.9.0",
-    "jest-extended": "^0.11.4",
-    "jest-html-reporter": "^2.8.0",
-    "jest-puppeteer": "^4.4.0",
     "lint-staged": "^10.2.11",
     "pkg": "^4.3.7",
     "prettier": "^1.19.1",
@@ -159,16 +152,14 @@
     "rimraf": "^2.6.2",
     "showdown": "^1.9.1",
     "stylus": "^0.54.5",
-    "ts-jest": "^24.3.0",
     "ts-node": "^9.1.1",
     "ts-node-dev": "^1.1.6",
     "typedoc": "^0.13.0",
     "typescript": "^3.9.7",
     "typescript-json-schema": "^0.32.0"
   },
-  "bin": "./out/index.js",
+  "bin": "../../out/nlu/index.js",
   "pkg": {
-    "assets": "../../build/native-extensions/**/*.node",
     "scripts": "out/**/*.js",
     "targets": [
       "host",
@@ -181,7 +172,6 @@
     "dev": "ts-node-dev ./index.ts",
     "start": "ts-node ./index.ts",
     "build": "tsc",
-    "package": "cross-env NATIVE_EXTENSIONS_DIR=./build/native-extensions pkg package.json",
-    "test": "cross-env NATIVE_EXTENSIONS_DIR=./build/native-extensions node_modules/.bin/jest -i --detectOpenHandles -c ../../jest.config.js ."
+    "package": "cross-env NATIVE_EXTENSIONS_DIR=./build/native-extensions pkg package.json"
   }
 }
